Add extension marker to i32 arguments of builtin functions
Some platform ABIs require i32 values to be zero- or sign-extended
to the full register width.  The extension is implemented by the
cranelift codegen backend, but this happens only if the appropriate
"uext" or "sext" attribute is present in the cranelift IR.

For calls to builtin functions, that IR is synthesized by the code
in func_environ.rs -- to ensure correct codegen for the target ABI,
this code needs to add those attributes as necessary.