Fix `branch_destination_mut()` accessor missing `IndirectJump`.
Found via fuzzing: the `branch_destination_mut()` accessor on
`InstructionData` was missing a case for `IndirectJump`. As per the
doc-comment, this should return `None`. Note that the `match` in
`branch_destination()` does have `IndirectJump` in the analogous place;
this appears to just have been an oversight.

Adding an LICM test from the fuzzer (minimized) because this is how it
was triggered; it's somewhat difficult to write a more direct test than
this.