Add module linking support to Wizer
This commit adds support for 99% of module linking to Wizer. This allows modules
within the "bundle" to import/export memories, globals, tables, and instances
from/to each other, but the root module still cannot import any external
state. The one big feature from module linking that isn't supported yet is
importing and exporting modules, rather than instances. More on this restriction
later.

Module linking breaks Wizer's previous assumption that a module is instantiated
and initialized exactly once (or, at least, that if it were instantiated and
initialized multiple times you would get the same result every time due to the
restrictions on imports and what can be used during initialization). Module
linking allows the same module to be instantiated multiple times and have a
different set of functions called on each instance during initialization. This
can lead to distinct initialized states for each instantiation. Here is a simple
example:

```wat
(module
  (module $A
    (global $g (mut i32) (i32.const 0))
    (func (export "f")
      global.get $g
      i32.const 1
      i32.add
      global.set $g))
  (instance $a1 (instantiate $A))
  (instance $a2 (instantiate $A))
  (func (export "wizer.initialize")
    call (func $a1 "f")
    call (func $a2 "f")
    call (func $a2 "f")))
```

After initialization, the `$a1` instance has had its `f` function export called
once and therefore its global's value is 1, while `$a2` has had its `f` function
export called twice and therefore its global's value is 2. Because each
instantiation has its own initialized state, we need to emit a pre-initialized
module for _each_ instantiation in the bundle.

At the same time, however, we do _not_ want to duplicate the `f` function across
each pre-initialized module! The code section is generally the largest section
in a Wasm module, and if we naively duplicated the code section for each
instantiation, we could get exponential code size blow ups. (For example,
consider a module linking bundle that contains a leaf module named `$Module1`
that is instantiated twice in another module called `$Module2`, which is in turn
instantiated twice in `$Module4`, which is in turn instantiated twice in
`$Module8`, etc... If the last module in this chain is instantiated `n` times,
then `$Module1` is instantiated `2^n` times, and if we duplicate the code
section for each instantiation we've got `2^n` copies of `$Module1`'s code
section.)

The solution is to split out a "code module" containing only the static sections
from the original module. All the state in the module (memories, globals, and
nested instantiations) are imported via a `__wizer_state` instance import. Then,
each instantiation defines its own "state module" which contains the specific
initialized state for that instance, and which is instantiated exactly once. To
create the initialized version of the original instantiation, we join the state
with the code by instantiating the code module and passing in this
instantiation's state instance.

Let's return to our original example from above. Here is the input Wasm again:

```wat
(module
  (module $A
    (global $g (mut i32) (i32.const 0))
    (func (export "f")
      global.get $g
      i32.const 1
      i32.add
      global.set $g))
  (instance $a1 (instantiate $A))
  (instance $a2 (instantiate $A))
  (func (export "wizer.initialize")
    call (func $a1 "f")
    call (func $a2 "f")
    call (func $a2 "f")))
```

And this is roughly what it looks like after wizening:

```wat
(module
  ;; The code module for $A.
  (module $A_code
    ;; Note that the code module imports state, rather than defining it.
    (import "__wizer_state"
      (instance (export "__wizer_global_0" (global $g (mut i32)))))
    (func (export "f")
      global.get $g
      i32.const 1
      i32.add
      global.set $g))

  ;; State module for $a1. Global is initialized to 1 since $a1's `f`
  ;; function was called once in the initialization.
  (module $a1_state_module
    (global (export "__wizer_global_0" (mut i32) (i32.const 1))))

  ;; The state instance for $a1.
  (instance $a1_state_instance (instantiate $a1_state_module))

  ;; Finally, we join the code+state together to create $a1.
  (instance $a1 (instantiate $A_code
                  (import "__wizer_state" $a1_state_instance)))

  ;; State module for $a2. Global is initialized to 2 since $a2's `f`
  ;; function was called twice in the initialization.
  (module $a2_state_module
    (global (export "__wizer_global_0" (mut i32) (i32.const 2))))

  ;; The state instance for $a2.
  (instance $a2_state_instance (instantiate $a2_state_module))

  ;; Finally, we join the code+state together to create $a2.
  (instance $a2 (instantiate $A_code
                  (import "__wizer_state" $a2_state_instance)))

  ;; Initialization function removed.
)
```

Using this code splitting technique allows us to avoid unnecessary duplication
for each instantiation. The only bits that are duplicated are the things that
are inherently different across instantiations: the instantiation's internal
state.

The reason we don't allow importing and exporting modules within the bundle is
that two different modules that have the same module type signature could have
two different sets of internal, not-exported state. This throws a wrench in the
instrumentation pass, which needs to force export all internal state, and in
this case could make the modules have incompatible type signatures, so they
couldn't both be imported by the same module anymore. That, in turn, would force
us to duplicate and specialize modules that import other modules for each
instantiation. This is do-able, but requires significant additional engineering
work. Therefore, this feature is left for some time in the future.

As a final note, adding support for module linking required pretty much a full
rewrite of Wizer. It no longer just parses the Wasm section by section, tweaking
it in place. Instead, there is an initial "parse" phase that creates a
`ModuleInfo` tree that serves as an AST. Our subsequent passes -- the
instrumentation and rewriting passes -- process the `ModuleInfo` tree rather
than reparsing the input Wasm and tweaking it directly.